Correction in specification for <label> position
The following thread details a problem with the specification, which states that a <label> element can only occur after the <name> and <et-al> elements within a <names> element. Following the specification would prevent users from controlling whether the label should be displayed before the name list or after the name list. Rintze answered that he probably made an error in drafting the specification, and that the CSL schema should be followed in this case - it allows the <label> to occur before and after the <name> element.

http://xbiblio-devel.2463403.n2.nabble.com/Names-label-position-td7578935.html